requirements.txt
asgiref==3.6.0
Django==4.2
django-stubs==1.16.0
django-stubs-ext==0.8.0
mypy==1.1.1
mypy-extensions==1.0.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.7.1.2
types-PyYAML==6.0.12.8
typing_extensions==4.5.0
djangorestframework==3.14.0